fix: modify code according to the latest comments
