+++ src/components/Chart/resolveLabelsOnBars.js
@@ -1,0 +1,26 @@
+/* eslint-disable no-underscore-dangle */
+import ChartJS from 'chart.js';
+import defaultTheme from '../../styles/defaultTheme';
+
+const palette = defaultTheme.palette;
+export default function resolveLabelsOnBars() {
+    const chart = this.chart;
+    const ctx = chart.ctx;
+
+    ctx.font = ChartJS.helpers.fontString(
+        ChartJS.defaults.global.defaultFontSize,
+        ChartJS.defaults.global.defaultFontStyle,
+        ChartJS.defaults.global.defaultFontFamily,
+    );
+    ctx.fillStyle = palette.text.label;
+    ctx.textAlign = 'center';
+    ctx.textBaseline = 'bottom';
+
+    this.data.datasets.forEach((dataset, i) => {
+        const meta = chart.controller.getDatasetMeta(i);
+        meta.data.forEach((bar, index) => {
+            const dataValue = dataset.data[index];
+            ctx.fillText(dataValue, bar._model.x, bar._model.y - 5);
+        });
+    });
+}
